Moved more functions to logger
